fix: refactor editProfileModal names, AddBelbinModalBody.jsx closes on submit, updated prop drilling in Import.jsx for edit modal
